extracted MissionDetailsStatus into its own file. added comments
